Make writing undefined clear persisted atom family values
Summary:
Previously we added localStorageBackedAtomFamily. We need a way to remove entries from local storage that are added, other than just waiting for them to get too old.

The use case is edited commit messages. If you start typing a message, it should be saved to local storage, so it appears when restarted. But if you actually amend/commit the message, we no longer need to persist it, and it's just taking up space in the storage. So ideally users of this atom can decide when to remove a value from storage.

The atomFamily is more convenient for users if you read values as `EditedMessage` and not `EditedMesssage | undefined`. However, we can simply make the derived atom's type allow writing `undefined`, but not reading `undefined`. This way, users who would like to remove an atom from storage can write `undefined` to it. This will actually use `getDefault(key)` as the value that gets stored in the atom, then tell the persisted storage to delete that key. This way, we don't need to worry about detecting if you try to write the default value or something like that. I basically inlined `atomWithOnChange` to accomplish this more easily.

One slight annoyance of this approach is that you need to explicitly use `undefined` to clear. You could imagine that a caller may write `{}` to editedCommitMessage(key), which happens to be exactly the default value. This would appear in the UI to have "nothing saved", but it would NOT delete from localstorage. I don't see a good way around this, and arguably it's good to have the control to be able to store the default value and remove from the cache separately.

Reviewed By: muirdm

Differential Revision: D61934556

fbshipit-source-id: de9851ae31f14153a54e08e196dc791f1965ed6c