Add localStorageBackedAtomFamily for editedCommitMessages
Summary:
Goal: Have commit messages you've typed but not amended/committed be persisted across closing and reopening ISL.

To do this, we need to store commit messages with our `platform` persistedStorage system. This uses localStorage in the browser and context.globalStorage in vscode.

Instead of a completely ad-hoc system, I opted to extend `atomFamilyWeak`. into `localStorageBackedAtomFamily`. The basic idea is that we have an atomFamilyWeak to store atoms as usual, and use a derived atom to convert writes into also storing the value in persisted storage.

Previous diffs in this stack ensured that the persisted storage system can track individual changes, so that only the one commit message being edited is sent on each update (as opposed to the entire persisted storage object).

In vscode, persisted storage values are included in the html so they are available synchronously. This could become problematic if commit messages are allowed to accumulate over months. So, the localStorageAtomFamily includes a maxAgeDays after which values are discarded on next load. This prevents unbounded accumulation of data. On startup, we need to read all values from persisted storage to filter out expired values by looking for the appropriate storage prefix. This is perhaps not fully ideal, but I think this is better than sending all persisted commit messages on every keystroke while typing a commit message.

In the next diff, we also make affordances to manually remove values from storage when they aren't needed, like when you commit/amend.

In the future, we could also add debouncing to the storage writes, so that we don't need to send messages on every keystroke.

Reviewed By: muirdm

Differential Revision: D61934555

fbshipit-source-id: 787afff6ad02d6fb53da09fa8f3991b7e064cea8