Don't send entire persisted storage data every time in vscode
Summary:
For persisted storage in vscode, we use vscode context.globalStorage, which is persisted on our behalf.

Previously, we had a single entry in this object called `isl-persisted`. This contained a stringified JSON value of all the smaller persisted key,value data.

This meant that anytime the webview changed any persisted data, it would send the entire set of persisted data to the extension. That was fine to start, but we plan on storing commit messages as you type them, which will become an issue if we have to send the entire state (aka all persisted commit messages) every keystroke.

The new approach is to use context.globalStorage with individual keys like `isl-persisted:has-shown-getting-started`.

On startup, we embed all persisted storage into the html (so that persisted data is synchronously available on webview startup). The equivalent here is to read through all globalState looking for 'isl-persisted:' prefixed keys, reading each of their data, and adding them to an object we JSON.stringify into the html.

(side note: despite the name, context.globalStorage is scoped to our extension ID. We do not see values stored by other extensions. The isl-persisted: prefix check is just future proofing in case we want to store something else in globalStorage unrelated to ISL)

Once we have the data on startup, any changes we make to any field now just sends the storage key and its relevant data, instead of all data. This leads to much smaller updates. Further reads are contained inside the webview without needing to go to the global storage again.

Reviewed By: muirdm

Differential Revision: D61887105

fbshipit-source-id: 63afad69562dfc96a0a9f784f8484f1bac9d50f0