Fix missing metrics, directly add successful fix stats
Summary:
# This Diff
- Fixes num_fix/fixable problem stats not being added
- Adds direct tracking of num_fixed_problem/problem_successful_fixes

# Why
Previously we did not directly track the successful fixes directly in the table because of table quota concerns and because we thought that it would be easy to make a derived column for that.
Since then we've come to realize eden isn't using that much scuba data compared to other uses so small size increases should be easy and that derived columns are hard, so these metrics are now in the table.

Differential Revision: D62052614

fbshipit-source-id: f3fff1271f26cfc0ed741a3dddf882f25ce1d4e5