Move the async_requests Thrift spec out of megarepo_api
Summary:
## This stack

Make `async_requests` general-purpose so that we can schedule more async work. The initial consumer will be SCS, but I expect this might be useful even more broadly than that.

## This diff

We'll need to make the request more generic, let's start by moving this out.

This was originally D61848170 but that made writers and workers incompatible, because we include the type name in Manifold keys. I have since landed D62024563 that makes this backward- and forward-compatible so it's now safe to re-apply.

Reviewed By: gustavoavena

Differential Revision: D62030834

fbshipit-source-id: 04c0afaa95bc59d5e1cf9058fd3a4986738d5186