Create a new crate for the async queue client
Summary:
## This stack

Make `async_requests` general-purpose so that we can schedule more async work. The initial consumer will be SCS, but I expect this might be useful even more broadly than that.

## This diff

`megarepo_api` had a lot of code that actually only exists to support clients and worker. Move it to a new crate to simplify the dependencies.

Reviewed By: gustavoavena

Differential Revision: D61907024

fbshipit-source-id: 958d862859ac87744447f90f1a96f9b910d5fed2