Remove verbose logging
Summary:
I added this in D61915516 to help me test the new code path; now I don't need it anymore.

Rather than a full revert, I'll only delete this log line so it's easier to add it back if needed. The compiler should smart enough that the rest of the diff in D61915516 should not matter.

Reviewed By: clara-9

Differential Revision: D62097682

fbshipit-source-id: 4cc752c283dc9257af47dcf64ebb39e59d18bc53