Move requests_table into async_requests
Summary:
## This stack

Make `async_requests` general-purpose so that we can schedule more async work. The initial consumer will be SCS, but I expect this might be useful even more broadly than that.

## This diff

The `requests_table` cargo only exists to support `async_requests` so it makes sense to move it there. I am not a fan of this crate-inside-a-crate, so I'll clean it up later in the stack; but this is a small improvement in unraveling from megarepo.

Reviewed By: gustavoavena

Differential Revision: D61850442

fbshipit-source-id: f05afbc4589992adaacf6a49860373595735178a