manifest: use comparison to implement find_intersection_of_diffs
Summary:
This function is a strict subset of `compare_manifest_tree`, including just the "added" and "changed" items, omitting the "same" and "removed" ones. Let's consolidate on the general implementation by reimplementing the old function in terms of it.

This also allows it to work on sharded manifests in a streaming fashion, and benchmarks show it is marginally (~5%) faster than the old implementation.

Differential Revision: D58675665

fbshipit-source-id: 52259e0dfc25157fcd28d6bc84878e025dbc7564