repo_factory: cache zelos clients based on config
Summary: Right now in `RepoFactory` we create a new zelos client whenever we create a new repo (e.g. during service startup or after config reloading) which requires creating a new connection to zelos. This diff makes it so that we use `RepoFactoryCache` instead to reuse zelos clients that were built using the same zelos config, similar to what we do for the blobstore and for xdb.

Reviewed By: andreacampi

Differential Revision: D62100691

fbshipit-source-id: 4cecafd8456d2da90fb9eccc8bd89f97992296c1