Fix mounting repo root showing everything as irrelevant
Summary:
A slight oversight here, if your cwd is the repo root, apparently the repo-relative cwd path is `/`. I'm not sure this is intended, but if that's true we should handle it properly in the irrrelevant cwd check. If you mount the repo root every commit should ALWAYS be relevant.

I thought I tested this but somehow it slipped through...

Reviewed By: zzl0

Differential Revision: D62132585

fbshipit-source-id: 9316efc2af0a768d3776221d4b05e066d587da1d