Allow Uncommit on a commit with children
Summary:
Previously, we only allowed Uncommit on the current commit `.` when it has no children. This is technically discouraged by `sl help uncommit`, but also `sl uncommit` actually runs just fine if done from a commit with children. It just leaves the original commit instead of hiding it. But you still move to the parent, and get the commit's changes as uncommitted changes.

Some people used this feature in the old ISL, so it seems reasonable to bring it back.

This requires updating our optimistic dag to understand this. We can also update our UI hints (tooltip and confirm message) to account for the difference.

I wonder if we should have `sl uncommit` mark the original commit as obsolete, so it's not "deleted", but the UI shows that it's no longer intended to exist. Though I could imagine some people not wanting that behavior either.

Reviewed By: quark-zju

Differential Revision: D55825886

fbshipit-source-id: 7216f85b548fa75d9ee1222c6a37818290afb678