Move OperatingSystem to components/
Summary:
This feels like a shared/ thing, but this file's only use is for isMac/isWindows, which snoop on userAgent—that's definitely only valid in the browser.

So, following our policy of having browser-only utils live in components/ instead of shared/, this diff moves this file over. This again reduces the dependency on shared/ from components/.

Reviewed By: quark-zju

Differential Revision: D62133724

fbshipit-source-id: 0a20ef1ae2268049e15f801a61730ee8677c59ae