tests: kill "sl()" bash function
Summary:
Similar to "hgmn()", this isn't needed anymore since the Sapling repos are configured properly.

This is a huge simplification (to me, at least), since now you just run "hg" instead of picking between "hg", "hgmn" and "sl", each with significant differences in Sapling configuration.

It was nice being able to use the name "sl", but the difference between "hg" and "sl" in the tests was confusing since "hg" and "sl" in real life are basically the same. In the future we should actually rename "hg" to "sl" in build and test.

Reviewed By: sggutier

Differential Revision: D62032629

fbshipit-source-id: 90590cb8e5d71851bca5f88f80327defeb2acee6