Ensure config hot reloading in used in BookmarkService
Summary: In S446520, we realized that `BookmarkService` did not get the latest version of the configuration even after the configerator diff landing and the latest config was applied only when the tasks restarted. With config hot reloading, this should not happen. The reason config hot reloading did not work is because the `BookmarkService` makes use of `MononokeRepos` instead of `MononokeReposManager`. This diff fixes that.

Reviewed By: YousefSalama

Differential Revision: D62178555

fbshipit-source-id: 5a2724bae17610d5ff6437297c906a102c660c08