get tab settings from VS Code
Summary:
When using the ISL extension in VS Code we don't really respect the tab size settings and always show them with a size of 8. This diff tries to fix that by getting the default tab size from VS Code when the ISL extension starts.

This is far from an ideal solution, however, since nowadays ISL starts as a webview, so we don't really update the settings after updating it. Additionally, this does does not respect more granularity (per language tab settings) either.

Reviewed By: evangrayk

Differential Revision: D62076263

fbshipit-source-id: 28adf94ea3891e6d9147ad645278cad787c6e3f0