Fix bookmark movement/infinite push interaction
Summary:
As seen in the parent diff, the intent is to bypass checks on additional changesets on bookmarks that match the infinitepush bookmark pattern.

Before this diff, it was achieved by not setting `affected_changesets` in `infinite_push_scratch_bookmarks` https://www.internalfb.com/code/fbsource/[26c867da7d3659d0468390109434c770ee703e71]/fbcode/eden/mononoke/repo_client/unbundle/src/processing.rs?lines=597, but that wouldn't work if `run_hooks_on_additional_changesets` was set to `true`, which has been the case in prod for a long time.

It would also not work with `bookmarks_movement_load_changesets_aggressive_simplification` which has been in prod for a shorter time.

Add an explicit override in affected changesets and respect it.

Reviewed By: singhsrb

Differential Revision: D62185104

fbshipit-source-id: 4c7c24030d100fc130f94606ff2fa3592a02b454