hooks: Use same settings as prod in tests and fix tests
Summary:
Since [a recent rollout](https://fb.workplace.com/groups/3112141815538647/permalink/8198791803540264/), prod is using the simplified logic for deciding when to run checks during bookmark movement.

The new behaviour on bookmark creation or move is like so:
* Run the checks (file hooks, changeset hooks, write permissions, ...) on all changesets that are being introduced with this bookmark move
* Run bookmark hooks once per bookmark move
* The jk: `scm/mononoke:run_hooks_on_additional_changesets` is never read (it is the default behaviour)
* The jk: `scm/mononoke:bookmarks_movement_use_precise_boundary` is never read (it is the default behaviour)
* The config: `hooks_skip_ancestors_of` is unused. No bookmark is special relative to this anymore
* The implementation config: `additional_changesets_limit` is unused. Instead, we don't potentially do unbounded work if the configs aren't exactly right

Many older tests were still testing the old behaviour which is on its way to being deprecated.

Update all integration tests and ensure that behaviours are consistent with the new expectations.
Some tests ended up shorter because there are now less edge-cases to test.

This updating of tests is a precursor to deleting all the code that became unused, and eventually all the jks and configs that are not used anymore.

Reviewed By: YousefSalama

Differential Revision: D62102404

fbshipit-source-id: 8eedac357c4f6fac1ad3a09ac3c74b45c69859e3