merge: fix backout copied file
Summary:
D61944971 tried to fix the backout issue of copied file, but the solution does
not cover the case when the working copy parent also changed the source file
of a copy.

This diff trys to fix it by detecting if a file was actually moved, only then
the file will be handled in the other side.

For changes in the tests:
* test-backout-rename.t: intended to fix by this diff
* test-glog.t: reverted changes of D60698882
* test-rebase-removed.t: seems fixed another previous bug (a bonus)
* test-rename-merge2.t: reverted changes of D60698882 and D55317450

Reviewed By: muirdm

Differential Revision: D62107144

fbshipit-source-id: 74cde701deaadbe331ad8a22ba50bed90f19c7be