Actually cache commit SLOC by hash
Summary:
I noticed some overfetching of SLOCs. The same hash was being requested 3x. This shouldn't be necessary, since SLOC will be constant per hash (changing files would change the hash). Thus, SLOC for a commit should be perfectly cacheable (not true for "pending amend" sloc).

Looking at the `commitSloc` atomFamily, I found it was not actually using the hash, and was instead taking the currently selected commit. This is weird and surprising that it even works. I guess since the only caller of this uses it in the commit info view for the head commit, it happens to line up to be the same data.

But more "correct" is to get the commit object corresponding to the hash that's requested—this way it doens't actually matter if you tried to get the SLOC for a commit that isn't being shown in the commit info view.

This happens to also remove the overfetching. I'm guessing a mismatch between the head commit info and the hash being used here was causing each newly selected hash to do multiple fetches before the cache started working. If it was supposed to follow that flow, it shouldn't have been an atom family, just a derived atom. I think the atomFamily is the better approach, as long as it's actually caching.

Reviewed By: benmonro

Differential Revision: D62197836

fbshipit-source-id: e1201fd215e8c354c5561eaee8722e26bfa97103