Fix duplicate tag problem in Mononoke Git Server
Summary: In our Mononoke Git Server, we made an implicit assumption that if two tags have different names then they would have different tag hashes. This is a logical assumption to make given that tags names become a part of the tag hash.  However, if certain operations are performed by hand on the server you can arrive at a situation where you have two tags with different names that have the same tag hash. In this case, due to the different names we would could two tags and send this count in the header of the packfile. However, when sending the actual contents of the packfile we won't be sending two objects but just one. This mismatch is what led to the clone failures of a bunch of AOSP repos and which is fixed by this diff.

Reviewed By: gustavoavena

Differential Revision: D62198325

fbshipit-source-id: 92d3de9a8860fd43bdbf13ba86f8e0139095a39a