Integration test highlighting the multiple tags with same tag hash problem
Summary: In our Mononoke Git Server, we made an implicit assumption that if two tags have different names then they would have different tag hashes. This is a logical assumption to make given that tags names become a part of the tag hash.  However, if certain operations are performed by hand on the server you can arrive at a situation where you have two tags with different names that have the same tag hash. In this case, due to the different names we would could two tags and send this count in the header of the packfile. However, when sending the actual contents of the packfile we won't be sending two objects but just one. This mismatch is what led to the clone failures of a bunch of AOSP repos and is reproduced in this integration test.

Reviewed By: gustavoavena

Differential Revision: D62197799

fbshipit-source-id: 99a85ad3152dd7610e0cbb4835aa1e0f9f9cad65