Fix bug in gitimport with missing inner tag
Summary:
In the scenario with a chain of annotated tags where some of them don't have a ref pointing at them (for instance, because it was deleted), gitimport must upload all the objects so serving the git protocol is even feasible.

Do this.

As a follow-up, we'll need to fix an issue in git server to allow us to serve `git clone`.

Reviewed By: YousefSalama

Differential Revision: D62237343

fbshipit-source-id: 8078e59d321645aa4a5d1b25aea3871fbc9e51ff