run hg debugnetwork --connection in eden doctor"
Summary:
# Context
Hg debugnetwork is a utility that diagnoses network issues. We already run this as part of eden rage. We now want to add this to eden doctor to add some additional signal on potential issues.

# This Diff
Add a call to hg debugnetwork --connection in eden doctor. This runs a check if eden can communicate with the remote server

# Technical Details
Debugnetwork --connection returns a variety of error codes to represent different issues. Rather than parse these, redirect the user to the tool since we can't really do any automatic fixing.

Reviewed By: kmancini

Differential Revision: D60774518

fbshipit-source-id: 05096ccfa78ab7f7ad0d38b4488b8128217445b2