run hg debugnetwork --speed in eden doctor
Summary:
# Context
Hg debugnetwork is a utility that diagnoses network issues. We already run this as part of eden rage. We now want to add this to eden doctor to add some additional signal on potential issues.

# This Diff
Add a call to hg debugnetwork --speed in eden doctor. This runs a check on the latency, download, and upload speed of the connection to the eden server

# Technical Details
The output is parsed to extract the latency, upload and download speed via regex.
Upload and download is always expressed in terms of Mbit/s
latency is a 4 significant figure number that is in either s, ms, or us

Reviewed By: kmancini

Differential Revision: D60774515

fbshipit-source-id: 78258a229e36470f718e63337003e543bcc55eb9