Implement privhelper pid request to server
Summary:
# Context
In D62051804, we add a feature that will remount hanging mounts upon eden reboot. We would like to have an integration test for this feature. One of the steps to create a hanging mount is to kill the privhelper. To do this we need the pid of the privhelper process.

# This Diff
Implement getServerPid.

# Technical Details
Standard server request-response implementation for pid_t.
The value is stored in the response immediately after the header and metadata.

Reviewed By: MichaelCuevas

Differential Revision: D62106989

fbshipit-source-id: bec7503cbfc4d84d41bd3f403848da84fc18cbfc