Add privhelper pid to PrivHelperInfo
Summary:
# Context
In D62051804, we add a feature that will remount hanging mounts upon eden reboot. We would like to have an integration test for this feature. One of the steps to create a hanging mount is to kill the privhelper. To do this we need the pid of the privhelper process.

# This Diff
Adds a getPid function to privhelper. Hook this up to the existing thrift endpoint checkPrivHelper.

# Technical Details
There are several ways a privhelper process is created. This method works if the privhelper is created when attemping to connect to the privhelper.
It does not work if the privhelper already exists and is connected to via socket. This will be added in the next diff.

Reviewed By: MichaelCuevas

Differential Revision: D62050113

fbshipit-source-id: 8a0b8d97be0cb5decfc9c4da30ea83fe994ded83