Fix for missing inner nested tag in Mononoke Git Server
Summary:
In D62197279, we demonstrated that git clone with the new git server can fail if there are nested tags in a repo and some refs corresponding to the tags in the nested chain are deleted from the repo. In this case, no direct references exist to the nested tags (via actual public refs) and they can only be reached by indirect pointers followed from the outermost tag.

With Pierre's fix in D62237343, we now import such orphaned nested tags and with my fix in this diff we would start serving them correctly in Mononoke Git server

Reviewed By: singhsrb

Differential Revision: D62264374

fbshipit-source-id: 8e4238ea1f3dbcf9a246c1cca931f992bbeecf66