Remove star rating UI
Summary:
This rating system looked nice but due to it's placement in the bug menu it didn't get used. Thus the results were statistically insignificant, widly changing from 5 star to 2 stars, etc. It had a  negative bias by being inside the bug menu people go to when something is wrong or they opt out.

Let's just remove this entirely. We have enough anecdotes of usefulness of ISL2 already, and this didn't provide measurable pulse to course correct with.

Reviewed By: sggutier

Differential Revision: D62255225

fbshipit-source-id: 2a6d0f3a27c1ee6766ed9d93d5aa7f055f517b3f