Split of "[Eden] Add Fix for: CheckoutConfig Corruption": pt2: move get_repo_info, RepoError, enbale_sqlite_overlay out of main.py
Summary:
# Context
Occasionally, config.toml gets corrupted or deleted. This file stores important information about the eden checkout.

# This Diff
Moves functions out of main.py to make them importable in init without a dependency loop

# Technical Details
- Creates a new CheckoutConfig based on info derived from the CheckoutInfo
- Backs up the old one
- Saves the new CheckoutConfig

# Caveats
This will only work on a currently running mount. If the mount is not running and config.toml is corrupted, we cannot determine what the backing repo is. In this case, we have return a EdenCheckoutInfoCorruption problem instead

Reviewed By: jdelliot

Differential Revision: D62304372

fbshipit-source-id: b5dabdc81d0b73a4c2d246177207219f4ecb2855