Add Fix for: CheckoutConfig Corruption pt.1 : config.py returns CheckoutConfigCoruptedError
Summary:
# Context
Occasionally, config.toml gets corrupted or deleted. This file stores important information about the eden checkout.

# This Diff
Adds a new FixableProblem EdenCheckoutConfigCorruption, which fixes the above issue by recreating config.toml using default values.

# Technical Details
- Creates a new CheckoutConfig based on info derived from the CheckoutInfo
- Backs up the old one
- Saves the new CheckoutConfig

# Caveats
This will only work on a currently running mount. If the mount is not running and config.toml is corrupted, we cannot determine what the backing repo is. In this case, we have return a EdenCheckoutInfoCorruption problem instead

Reviewed By: MichaelCuevas

Differential Revision: D61882849

fbshipit-source-id: b8d17c9e9d4b08bbefc8e1117b9128122f582cae