fbcodereview: fix global flag passing to marklanded hook
Summary:
I noticed that `-v` no longer prints what commits are marked as landed.
Fix it by passing the global flags to the hook command.

As we're here, use `util.gethgcmd()` to use the current program so `hg` won't
refer to the system hg in PATH. i18n should not be used to translate
`debugmarklanded`.

Reviewed By: muirdm

Differential Revision: D62308319

fbshipit-source-id: 438a6797e66608a9f3a6f5cceba3628656367116