Increase commit processing performance by 4x
Summary:
The commit processing and uploading pipeline used by `gitimport`, `remote-gitimport` and `Mononoke Git server` is not as efficient as we would like it to be. In particular, there is unnecessary sequentialization in the initial part of the pipeline that prevents us from using the available concurrency on the server.

After we get the set of commits to process, we follow these steps:
1. Fetch the necessary metadata related to the Git commit
2. Find the file (blob) changes introduced by that commit
3. Upload the blob changes, LFS content and packfile item for the blobs
4. Identify the directory (tree) changes introduced by that commit
5. Upload the tree changes and packfile item for the trees
6. Upload the git commit and its associated packfile item
7. Generate the bonsai changeset for the git commit
8. Store the bonsai changeset (and its mapping to git commit) in storage and derive the necessary derived data

Out of all these steps, only steps 7 and 8 need to happen purely sequentially otherwise we get unsoundness. We already decoupled step 7 and 8 in order to speed things up but it wasn't sufficient. This diff goes a step further and does steps 1 - 6 concurrently and only then introduces sequential ordering. Instead of using a single stream connected by `.buffered()` variants, I used `mpsc` channel to ensure that independent streams can make progress independently. Now that we are leveraging better concurrency, I increased the default concurrency value for Mononoke Git Server

Additionally, in this diff I fixed the reporting of progess using a dedicated counter.

Differential Revision: D62356435

fbshipit-source-id: 39a39a6e7446895f773a79cb2f6bc83eaa8617c3