Enable deletion of configurable folders as part of `redirect fixup`(3/n): Enable force deletion of well known non-empty folders in `eden doctor` flow
Summary:
## Background:
As part of D60802870, we improved the messaging to make error message crisper and actionable in case of redirection failures because of non-empty folders. But still we expect users to pass `--force` flag to ensure that they are deleting the content of a folder which might have non-recoverable changes. This still needs fiddling around a little to understand what's going on.

## Solution
We know that most of these redirected folders, like buck-out, are created by tools for temporary purpose for keeping artifacts which are built or generated as part of development process and can be re-generated again without losing any work. Proposal is to keep a configurable list of paths which could be deleted without user consent and delete it as part of `eden redirect fixup` without `--force` flag.

# Implementation:
1. [D60613326] Introduce `redirect-fixup-deletable-paths` config in Configerator and initialize it with some paths from [`.eden-redirections` files](https://www.internalfb.com/code/search?nl_enabled=false&q=repo%3Aall%20.eden-redirections&tab=path).
2. Using this config, force delete these paths in rust script
3. **Using this config, force delete these paths in python script**

Reviewed By: genevievehelsel

Differential Revision: D62194080

fbshipit-source-id: a4e04574cb52f11d565171330593b9da573ff287