add dialoguer as a dependency for linux and windows
Summary:
`dialoguer` is now supported with Windows and Linux. We should migrate to it instead of implementing the prompt code ourselves.

This diff marks dialoguer as a general dependency instead of os-specific ones.

Note: `dialoguer` is behaving differently from the Python prompt. As long as "Y/y" or "N/n" is typed it will take it as an input (no "enter" required). this should be fine if we keep the same behavior across Rust commands on all platforms.

Reviewed By: MichaelCuevas

Differential Revision: D62274190

fbshipit-source-id: 843a64ba14a3ef9c720ebdae7a1b74974deacbb8