config: allow typed config conversions to access config
Summary:
Add a new FromConfig trait which is similar to FromConfigValue but also gets the `&dyn Config` object itself.

This is so the type can look up other config values as needed when parsing the config value. I'm going to use this for parsing "paths" in the config into a RepoUrl object.

Reviewed By: zzl0

Differential Revision: D62300483

fbshipit-source-id: a8cf978bc478b99b16b302ca4715afd0bd0ba941