repo: resolve path schemes when making SLAPI client
Summary:
When trying to construct an SaplingAPI client based on paths.default, we weren't resolving custom schemes.

The Mononoke tests have a "mono=ssh://user@dummy/{1}" scheme that needs to be resolved so the eagerepo code looking for "ssh://user@dummy" works.

Fix by using the RepoUrl object in more spots (which resolves the schemes for you).

Reviewed By: andreacampi

Differential Revision: D62300482

fbshipit-source-id: 339ff1ae5eb156065edeb22296605adb8b5affad