implement digestHash fetching
Summary:
# Context:

We are currently overloading the pre-existing attributes `FILE_SIZE` and `BLAKE3` to mean different things depending on whether directory or file attributes are being fetched from the readdir/getAttributesFromFilesV2 endpoint.

In the file case, file_size means the size of the file on disk. In the directory case, it means the size of the directory digest that's stored in RE's Content Addressed Store (CAS).

In the file case, blake3 means the blake3 hash of that single file. In the directory case, blake3 means the blake3 hash of the directory digest that's stored in RE CAS.

It's no longer fine to overload the blake3 hash case since existing users of the blake3 attribute do NOT expect the number of results to increase. Therefore, we will have to introduce a separate attribute for requesting hashes of directories that won't interfere with the current usage of blake3 attribute requests.

# Solution

We should introduce a new attribute for query RE CAS digest size information for directories. This will allow us to avoid overloading the existing size/blake3 attributes and instead allow users to specifically query for CAS digest size/hash if they want it. This attribute will be queriable for both directories and files. However, for files it will behave the exact same way as requesting the existing file_size/blake3 attribute.

# This diff

implements digest hash fetching for trees and files. For files, we simply reuse the existing blake3 attribute result. For trees, we replace the blake3 attribute result w/ this new attribute to avoid overloading existing attributes.

Reviewed By: jdelliot

Differential Revision: D61689073

fbshipit-source-id: 8d39629466c1c8cd3f298fc581b9bca1e848c1f5