Make "next to run" collapsable
Summary: Add a disclosure chevron and state to make this expandable or not. This way you can hide this list if it gets really long and is in your way. Ideally commands go quickly and this isn't necessary. On small screens, it's nice to have that control over the UI.

Reviewed By: quark-zju

Differential Revision: D62393043

fbshipit-source-id: d7a20d975ebb325b8f82fdc3cd4077091adfc515