revisionstore: remove ContentStore
Summary:
This store was superseded by ScmStore. It hasn't been used in production in a long time, but hung around for tests because it was the only Rust store that could read remote data over the wireproto (using Python). Now that SLAPI should be available in all tests, ScmStore is sufficient and we don't need ContentStore.

I moved a few minor bits of code out of contentstore since they were used elsewhere.

Reviewed By: quark-zju

Differential Revision: D62337841

fbshipit-source-id: 05d2cc46efe8d8b41c5d18aa697d751067ad12cd