When listing requests, include any that haven't started
Summary:
The previous implementation only showed requests once a worker started processing them. Until then, the `inprogress_last_updated_at` column is empty and would be skipped.

That is clearly not intentional, and it's confusing. In practice, I think it never came up because in normal operation a job is dequeued immediately; but it is now showing up in integration tests.

This query is only invoked by the CLI so it's safe to change. If there are use cases where we legitimately don't want to see new entries, we should add filtering to the CLI.

Reviewed By: mitrandir77

Differential Revision: D62383330

fbshipit-source-id: 079eb35b4a39b07942ae8d3e8203ee102f03a330