screenshot: initial scaffolding to script ISL recordings
Summary:
Use Puppeteer (the `-core` variant uses the system broswer instead of bundling
its own) to take screenshot or recording.

This diff adds basic scaffolding to prepare a repo, launch ISL, open the
browser, and perform some custom actions. The "prepare repo" and
"custom actions" parts are abstracted in `Example` to support multiple
examples.

To run an example:

  yarn start exampleName  # "screenshot" or "repl"

Dependencies are intended to match other ISL projects, except for `execa` which
can only be imported as ESM in newer versions. Since the dependency tree is
different, I didn't make it share the workspace for now.

This is far away from a complete recording tool. Namely, miss the following:
- Prepare a repo with a "remote" and "reasonable" draft commits.
- Recording. Show the mouse cursor during recording.

Nonetheless, it can take a screenshot just fine and can start an REPL
to play with the Puppeteer APIs.

Reviewed By: evangrayk

Differential Revision: D62330871

fbshipit-source-id: 079a3fd5774416decbb9bbbd917804acb29140df