cas: introduce new config options that would allow to enable files only or
Summary:
introduce new config options that would allow to enable files only or
trees only in CAS. This is easy to do because separate instances of CAS client
are used for trees and for files. The purpose of this change is to de-risk
rollout. The plan is to start initially with blobs only from CAS. This should also help with performance testing, so that we can start with blobs.

Reviewed By: muirdm

Differential Revision: D62382579

fbshipit-source-id: 0754789e6fd6c0ab3f5b0d403f2a623b210268a1