add a way to check permissions to create repos
Summary:
Before we can add an endpoint to create repos we need to have a way to do ACL
check.

Context: This stack adds new SCS endpoint to create git repos. For now
we'll start with Metagit repos but we're aiming to make the API future-proof
and ready for creating Mononoke repos.

Reviewed By: lmvasquezg

Differential Revision: D62393087

fbshipit-source-id: f851094c586a6c641f8b8defb0f5dee2747068e3