test endpint for repo creation
Summary:
SCSC command and interation tests allow us to at least test if we call the
scmadmin command and do the ACL checks. The actual repo creation can't be
fully tested outside of prod.

Context: This stack adds new SCS endpoint to create git repos. For now
we'll start with Metagit repos but we're aiming to make the API future-proof
and ready for creating Mononoke repos.

Reviewed By: gustavoavena

Differential Revision: D62393085

fbshipit-source-id: eeb723546a45a68423784fa61247ccb967eb32c2