Support allow-listing specific binaries by sha256 and size
Summary:
See [this post](https://fb.workplace.com/groups/p4pqa/posts/8243379959049980) and its comments for more context.

They want to be able to allow-list a specific executable in every path.

## Note about the Sha256 type in the config

I initially defined the type in the config as a `mononoke_types::hash::Sha256`. But as it turns out, it would be a huge pain to use it in the hooks config.
I discussed it with markbt and I could

1) Spend **a lot more time** than I planned implementing proper serialization/deserialization to pass the Sha256 as a string in the config.
2) Pass the sha as a bytes of `u8`, which would work but might lead to someone misusing in the future and breaking the services when trying to load the hook configs.
3) Define the sha256 as string.

(3) is not great, but it's the one I deemed best for the time I have right now.

Reviewed By: mitrandir77

Differential Revision: D62429882

fbshipit-source-id: 2506ad43ba2c18a153efadba549ab8f691dc020e