mononoke_types: don't derive serde ops for the hash types
Summary:
The hash types currently derive serde `Serialize` and `Deserialize`.  However, since they are wrappers around `[u8; N]`, the default implementation uses a list of integers.  This is inefficient and also produces confusing views of the hashes, whereas normally you'd expect to see a hex string.

Prevent this foot-gun by removing the derived operations.  As it happens, this is only used in one place (a test tool) and nothing actually makes use of that.

Later, if we want to serialize these types, we should implement the traits directly and store them as an array of bytes or hex-encoded (depending on the target).

Reviewed By: mitrandir77

Differential Revision: D62438677

fbshipit-source-id: 7025c1a40813ca4883e9b1c0d8c92f673409781e