treemanifest: kill treemanifestserver extension
Summary:
This extension is a pain in the butt. It duplicates a lot of code with treemanifest (which is confusing for developers), and is incompatible with debugruntest in-process execution since it conflicts with the treemanifest extension in various ways.

I think maybe we don't need treemanifestserver anymore, so let's try killing. I am fuzzy on the details, but we should be using modern stores in all cases for trees now, and we don't need to support the flat->tree transition, so maybe the regular treemanifest extension can work for both client and server.

I got rid of some treemanifest config options:
 - "treemanifest.treeonly" is always true (if treemanifest extension is loaded)
 - "treemanifest.server" is not needed now that client and server use the same storage layers
 - "treemanifest.rustmanifest" is always true
 - "treemanifest.sendtrees" is always true
 - "treemanifest.useruststore" is always true
 - "treemanifest.flatcompat" is always false

Reviewed By: quark-zju

Differential Revision: D54889477

fbshipit-source-id: c881d98d2d4de9607b174790e1cf166dc404bd44